[Prover Service] Merge verified input type into request handler types.
