+++ prover-service/src/input_processing/mod.rs
@@ -5,4 +5,3 @@ pub mod field_parser;
 pub mod input_signals;
 pub mod public_inputs_hash;
 pub mod rsa;
-pub mod types;
