+++ prover-service/src/training_wheels/mod.rs
@@ -6,9 +6,8 @@ pub mod verification_logic;
 use crate::external_resources::jwk_fetcher;
 use crate::external_resources::jwk_fetcher::{get_federated_jwk, JWKCache};
 use crate::external_resources::prover_config::ProverServiceConfig;
-use crate::input_processing::types::VerifiedInput;
 use crate::request_handler::prover_state::ProverServiceState;
-use crate::request_handler::types::RequestInput;
+use crate::request_handler::types::{RequestInput, VerifiedInput};
 use crate::training_wheels::verification_logic::compute_nonce;
 use anyhow::{anyhow, bail, ensure};
 use aptos_keyless_common::input_processing::encoding::{AsFr, DecodedJWT};
